+++ src/main/kotlin/uk/co/ben_gibson/git/link/ui/settings/DomainRegistrySettings.kt
@@ -9,9 +9,9 @@ import com.intellij.ui.layout.panel
 import com.intellij.ui.table.TableView
 import com.intellij.util.ui.ColumnInfo
 import com.intellij.util.ui.ListTableModel
-import uk.co.ben_gibson.git.link.GitLinkBundle
 import uk.co.ben_gibson.git.link.GitLinkBundle.message
 import uk.co.ben_gibson.git.link.git.Host
+import uk.co.ben_gibson.git.link.git.Hosts
 import uk.co.ben_gibson.git.link.git.HostsProvider
 import uk.co.ben_gibson.git.link.settings.ApplicationSettings
 import uk.co.ben_gibson.git.link.ui.components.HostCellRenderer
@@ -21,13 +21,12 @@ import uk.co.ben_gibson.git.link.ui.validation.domain
 import uk.co.ben_gibson.git.link.ui.validation.exists
 import uk.co.ben_gibson.git.link.ui.validation.notBlank
 import java.awt.event.ItemEvent
-import java.net.URI
 import javax.swing.ListSelectionModel
 
-class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.auto-detect.group.title")) {
+class DomainRegistrySettings : BoundConfigurable(message("settings.domain-registry.group.title")) {
     private val settings = service<ApplicationSettings>();
     private val hosts = service<HostsProvider>().provide();
-    private var customDomains = settings.customHostDomains
+    private var domainRegistry = settings.customHostDomains
 
     private val hostsComboBoxModel = HostComboBoxModelProvider.provide()
 
@@ -36,7 +35,7 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
     private val domainsTable = TableView(domainsTableModel).apply {
         setShowColumns(true)
         setSelectionMode(ListSelectionModel.SINGLE_SELECTION)
-        emptyText.text = message("settings.auto-detect.table.empty")
+        emptyText.text = message("settings.domain-registry.table.empty")
     }
 
     private val domainsTableContainer = ToolbarDecorator.createDecorator(domainsTable)
@@ -73,11 +72,11 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
 
     private fun addDomain() {
         val host = hostsComboBoxModel.selected ?: return
-        val dialog = RegisterDomainDialog(existing = customDomains.getOrDefault(host.id.toString(), setOf()))
+        val dialog = RegisterDomainDialog(domainsRegistry = domainRegistry, hosts = hosts)
 
         if (dialog.showAndGet()) {
-            val existingUrls = customDomains.getOrDefault(host.id.toString(), setOf())
-            customDomains = customDomains.plus(Pair(host.id.toString(), existingUrls.plus(dialog.domain)))
+            val hostDomains = domainRegistry.getOrDefault(host.id.toString(), setOf())
+            domainRegistry = domainRegistry.plus(Pair(host.id.toString(), hostDomains.plus(dialog.domain)))
             refreshDomainsTable(host)
         }
     }
@@ -86,9 +85,9 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
         val host = hostsComboBoxModel.selected ?: return
         val remove = domainsTable.selectedObject ?: return
 
-        val existingUrls = customDomains.getOrDefault(host.id.toString(), setOf())
+        val hostDomains = domainRegistry.getOrDefault(host.id.toString(), setOf())
 
-        customDomains = customDomains.plus(Pair(host.id.toString(), existingUrls.minus(remove)))
+        domainRegistry = domainRegistry.plus(Pair(host.id.toString(), hostDomains.minus(remove)))
 
         refreshDomainsTable(host)
     }
@@ -97,19 +96,19 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
         val host = hostsComboBoxModel.selected ?: return
         val domain = domainsTable.selectedObject ?: return
 
-        val dialog = RegisterDomainDialog(domain, customDomains.getOrDefault(host.id.toString(), setOf()))
+        val dialog = RegisterDomainDialog(domain, domainRegistry, hosts)
 
         if (dialog.showAndGet()) {
-            val existingUrls = customDomains.getOrDefault(host.id.toString(), setOf())
+            val existingUrls = domainRegistry.getOrDefault(host.id.toString(), setOf())
 
-            customDomains = customDomains.plus(Pair(host.id.toString(), existingUrls.minus(domain).plus(dialog.domain)))
+            domainRegistry = domainRegistry.plus(Pair(host.id.toString(), existingUrls.minus(domain).plus(dialog.domain)))
 
             refreshDomainsTable(host)
         }
     }
 
     private fun refreshDomainsTable(host: Host) {
-        domainsTableModel.items = host.domains.map { it.toString() }.toList() + customDomains.getOrDefault(host.id.toString(), setOf())
+        domainsTableModel.items = host.domains.map { it.toString() }.toList() + domainRegistry.getOrDefault(host.id.toString(), setOf())
     }
 
     private fun createDomainsTableModel(domains: List<String> = listOf()): ListTableModel<String> = ListTableModel(
@@ -126,7 +125,7 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
     override fun reset() {
         super.reset()
 
-        customDomains = settings.customHostDomains
+        domainRegistry = settings.customHostDomains
 
         val host = hostsComboBoxModel.selected ?: return
 
@@ -134,17 +133,21 @@ class AutoDetectSettings : BoundConfigurable(GitLinkBundle.message("settings.aut
     }
 
     override fun isModified() : Boolean {
-        return super.isModified() || customDomains != settings.customHostDomains
+        return super.isModified() || domainRegistry != settings.customHostDomains
     }
 
     override fun apply() {
         super.apply()
 
-        settings.customHostDomains = customDomains
+        settings.customHostDomains = domainRegistry
     }
 }
 
-class RegisterDomainDialog(var domain: String = "", private val existing: Set<String>) : DialogWrapper(false) {
+class RegisterDomainDialog(
+    var domain: String = "",
+    private val domainsRegistry: Map<String, Set<String>>,
+    private val hosts: Hosts
+) : DialogWrapper(false) {
     init {
         title = message("settings.auto-detect.register-domain-dialog.title")
         setOKButtonText(if (domain.isEmpty()) { message("actions.register") } else { message("actions.update") })
@@ -156,7 +159,13 @@ class RegisterDomainDialog(var domain: String = "", private val existing: Set<St
         row(message("settings.auto-detect.register-domain-dialog.title")) {
             textField(::domain)
                 .focused()
-                .withValidationOnApply { notBlank(it.text) ?: domain(it.text) ?: exists(it.text, existing ) }
+                .withValidationOnApply {
+                    notBlank(it.text) ?:
+                    domain(it.text) ?:
+                    exists(it.text, hosts.toSet().flatMap { host -> host.domains.map { domain -> domain.toString() } } ) ?:
+                    exists(it.text, domainsRegistry.values.flatten())
+                }
+
         }
     }
 }